* PRTBND-1099 Fixed Update Sales Person Failing if Email Not Provided

+++ app/Services/CRM/User/SalesPersonService.php
@@ -91,13 +91,12 @@ class SalesPersonService implements SalesPersonServiceInterface
         $params = $this->mergeSmtpImap($rawParams);
 
         // Find Existing Sales Person Email On a DIFFERENT Sales Person
-        $existing = $this->salespeople->getByEmail($params['user_id'], $params['email'], $params['id']);
-        if(!empty($existing->id)) {
-            // Delete current sales person and update matching one instead if it is deleted
-            if(!empty($existing->deleted_at)) {
+        if(!empty($params['email'])) {
+            $existing = $this->salespeople->getByEmail($params['user_id'], $params['email'], $params['id']);
+            if(!empty($existing->id) && !empty($existing->deleted_at)) {
                 $this->salespeople->delete(['id' => $params['id']]);
                 $params['id'] = $existing->id;
-            } else {
+            } elseif(!empty($existing->id)) {
                 // Throw exception instead!
                 throw new DuplicateChangeEmailSalesPersonException;
             }
